Add test_lobby_communique.py
Move communique tests to a new file.
